init redux and move `App` to components folder
