+++ frontend/src/store/index.ts
@@ -1,0 +1,17 @@
+import { configureStore } from '@reduxjs/toolkit';
+import appSlice from 'components/App';
+import gridSlice from 'components/Grid/gridSlice';
+import questionsSlice from 'components/QuestionPanel/questionsSlice';
+
+const store = configureStore({
+  reducer: {
+    app: appSlice,
+    grid: gridSlice,
+    questions: questionsSlice,
+  },
+});
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
+
+export default store;
